+++ BGL/include/CGAL/boost/parameter.h
@@ -31,9 +31,8 @@
 #else
 #  define  BOOST_PARAMETER_MAX_ARITY 12
 #endif
-#include <boost/parameter.hpp>
-#include <boost/parameter/name.hpp>
 
+#include <boost/parameter/name.hpp>
 
 #if defined(__clang__) || (BOOST_GCC >= 40600)
 #  define CGAL_IGNORE_UNUSED_VARIABLES \
@@ -58,7 +57,7 @@ namespace CGAL
 {
 namespace parameters
 {
-  
+
 template <typename T>
 struct Base
 {
@@ -113,6 +112,20 @@ BOOST_PARAMETER_NAME( (dump_after_exude_prefix, tag ) dump_after_exude_prefix_)
 BOOST_PARAMETER_NAME( (number_of_initial_points, tag) number_of_initial_points_)
 BOOST_PARAMETER_NAME( (maximal_number_of_vertices, tag ) maximal_number_of_vertices_)
 BOOST_PARAMETER_NAME( (pointer_to_error_code, tag ) pointer_to_error_code_)
+
+// First used in <CGAL/Labeled_mesh_domain_3.h>
+BOOST_PARAMETER_NAME( (function, tag ) function_)
+BOOST_PARAMETER_NAME( (bounding_object, tag ) bounding_object_)
+BOOST_PARAMETER_NAME( (relative_error_bound, tag ) relative_error_bound_)
+BOOST_PARAMETER_NAME( (p_rng, tag ) p_rng_)
+BOOST_PARAMETER_NAME( (null_subdomain_index, tag ) null_subdomain_index_)
+BOOST_PARAMETER_NAME( (construct_surface_patch_index, tag ) construct_surface_patch_index_)
+
+// First used in <CGAL/Gray_image_mesh_domain_3.h>
+BOOST_PARAMETER_NAME( (image, tag ) image_)
+BOOST_PARAMETER_NAME( (iso_value, tag) iso_value_)
+BOOST_PARAMETER_NAME( (value_outside, tag) value_outside_)
+BOOST_PARAMETER_NAME( (image_values_to_subdomain_indices, tag ) image_values_to_subdomain_indices_)
 
 CGAL_PRAGMA_DIAG_POP
 } // parameters