+++ Mesh_3/examples/Mesh_3/mesh_3D_gray_image_multiple_values.cpp
@@ -8,7 +8,10 @@
 #include <CGAL/Labeled_mesh_domain_3.h>
 #include <CGAL/make_mesh_3.h>
 #include <CGAL/Image_3.h>
-#include <functional>
+#include <cstdlib>
+#include <iostream>
+
+#include <boost/container/flat_set.hpp>
 
 typedef float Image_word_type;
 
@@ -26,6 +29,16 @@ typedef CGAL::Mesh_criteria_3<Tr> Mesh_criteria;
 // To avoid verbose function and named parameters call
 using namespace CGAL::parameters;
 
+template <typename Set>
+struct Image_to_multiple_iso_level_sets {
+  const Set& set;
+  Image_to_multiple_iso_level_sets(const Set& set) : set(set) {}
+
+  int operator()(float v) const {
+       return int(std::distance(set.begin(), set.lower_bound(v)));
+  }
+};
+
 int main(int argc, char*argv[])
 {
   const char* fname = (argc>1)?argv[1]:"data/skull_2.9.inr";
@@ -35,10 +48,26 @@ int main(int argc, char*argv[])
     std::cerr << "Error: Cannot read file " <<  fname << std::endl;
     return EXIT_FAILURE;
   }
-  /// [Domain creation]
+
+  typedef boost::container::flat_set<float> Flat_set;
+  Flat_set iso_values;
+  if(argc < 3) {
+    iso_values.insert(2.9f);
+  } else {
+    for(int i = 2; i < argc; ++i) {
+      iso_values.insert(static_cast<float>(std::atof(argv[i])));
+    }
+  }
+
+  // Domain
+  namespace p = CGAL::parameters;
   Mesh_domain domain =
-    Mesh_domain::create_gray_image_mesh_domain(image, 2.9f, 0.f);
-  /// [Domain creation]
+    Mesh_domain::create_gray_image_mesh_domain
+    (p::image = image,
+     p::image_values_to_subdomain_indices =
+       Image_to_multiple_iso_level_sets<Flat_set>(iso_values),
+     p::value_outside = 0.f
+     );
 
   // Mesh criteria
   Mesh_criteria criteria(facet_angle=30, facet_size=6, facet_distance=2,