+++ Mesh_3/include/CGAL/internal/Mesh_3/helpers.h
@@ -119,12 +119,10 @@ struct Is_border_edge {
   }
 }; // end Is_featured_edge<Polyhedron>
 
-template<typename Polyhedral_mesh_domain,
-         typename Polyline_with_context,
+template<typename Polyline_with_context,
          typename Graph>
 struct Extract_polyline_with_context_visitor
 {
-  typedef typename Polyhedral_mesh_domain::Polyhedron_type Polyhedron;
   std::vector<Polyline_with_context>& polylines;
   const Graph& graph;